asjaskdfh

+++ src/main/java/com/yole/carapp/exceptions/ResourceNotFoundException.java
@@ -1,7 +1,7 @@
 package com.yole.carapp.exceptions;
 
-public class ResourceNotFoundException extends RuntimeException{
-    public ResourceNotFoundException(String resource, int id) {
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String resource, Long id) {
         super(resource + " with ID " + id + " not found.");
     }
 }
